Overview: Milestone 2 review - Added in functions for printing each of the player's four tokens to the path object: need to work on centering algorithms.
The entirety of Milestone 2 has been met: we've reached our self-made goal of presenting the board in the proper state, as well as creating the basis of the game's finite state machine. The process of the finite state machine is handled by a static StateManager class where as values and information are passed down the chain to specific portions of the UI and token objects.

We reorganized the UI to handle the data in a fashion that allows (x,y) coordinate calculations to take place far before a print statement is called, allowing for us to anticipate the placement of the button objects far before they are printed to the screen. More states (such as animation states) are in the process of being added to the World manager and the Player classes. This is to allow the player classes the necessary "states". Players shouldn't be able to roll, for example, during their idle phase or on another player's turn.

This is all being tied together by the World manager's class - our process consists of building functions directly into the Player class and then working them outwards into the World class when it comes to the state machines. Other functions such as data from the Board class are pushed through the World manager. Rather, it's more accurate to say that the World manager pulls data from the Board and Player classes to find the proper reference to the right Path object, validating these values, making on the spot calculations (for things such as center positions) and then passing it to the Player's draw method for final printing. The problem lies with the ambiguous nature of finding an object's center with any object in XNA. We're working to make a central calculation that will take a series of x, y, width, and height parameters, and thusly return the boundaries that should be used to print to the screen.