Overview: NOT COMPLETE - Token Animation
Detailed: Most of the functionality has been programmed. The tokens have an off by one error currently (should be able to fix it but I don't have the time right now). The tokens functionally move (but they don't animate, don't know why yet) and they move all along the board (they break on a full loop though because the GetPath function is broken). Most of the logic is there though. Fix the off by one error, animation not being smooth error, and GetPath function and we are good to go.