Overview: Implemented cards having cost.
Detailed: Implemented loading cards and decks with cost. Also implemented a new messagebox that asks the player if they want to buy first, and if so, the card is displayed and they lose money. Only problem currently is that the drawing of the message box is still wonky. I don't want to fix this until we have a final font because it will change when we change the font.